test_requirements.txt
# Testing Framework Requirements
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0
pytest-xdist>=3.0.0
pytest-html>=3.1.0
pytest-mock>=3.10.0

# Browser Testing
selenium>=4.15.0
webdriver-manager>=4.0.0

# Visual Testing
Pillow>=10.0.0

# Performance Testing
psutil>=5.9.0
memory-profiler>=0.61.0

# Accessibility Testing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# API Testing
requests>=2.31.0
responses>=0.23.0

# Code Quality
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0

# Coverage Reporting
coverage>=7.0.0
coverage-badge>=1.1.0